backend/clustering/requirements.txt
fastapi
uvicorn
requests
mysql-connector-python
pandas
schedule
pyyaml
openpyxl
scikit-learn